Add files via upload
We tested the association between the PRS and the behavioral manifestation of the trait.  Associations between PRS-ADHD across all thresholds and behavioral measures were estimated using linear mixed-eﬀects models with school as nested random effect to account for the multilevel nature of the data (i.e. children within schools).